Cypress 10 and custom commands are too complicated
