--- client/cypress/support/commands.ts
@@ -1,83 +1,0 @@
-// ***********************************************
-// This example commands.js shows you how to
-// create various custom commands and overwrite
-// existing commands.
-//
-// For more comprehensive examples of custom
-// commands please read more here:
-// https://on.cypress.io/custom-commands
-// ***********************************************
-//
-//
-// -- This is a parent command --
-// Cypress.Commands.add("login", (email, password) => { ... })
-//
-//
-// -- This is a child command --
-// Cypress.Commands.add("drag", { prevSubject: 'element'}, (subject, options) => { ... })
-//
-//
-// -- This is a dual command --
-// Cypress.Commands.add("dismiss", { prevSubject: 'optional'}, (subject, options) => { ... })
-//
-//
-// -- This will overwrite an existing command --
-// Cypress.Commands.overwrite("visit", (originalFn, url, options) => { ... })
-
-const COMMAND_DELAY = 300;
-
-for (const command of [
-  "visit",
-  "click",
-  "trigger",
-  "type",
-  "clear",
-  "reload",
-  "contains",
-  "submit"
-]) {
-  Cypress.Commands.overwrite(command, (originalFn, ...args) => {
-    const origVal = originalFn(...args);
-
-    return new Promise((resolve) => {
-      setTimeout(() => {
-        resolve(origVal);
-      }, COMMAND_DELAY);
-    });
-  });
-}
-
-declare global {
-  namespace Cypress {
-    interface Chainable {
-      /**
-       * Provides a working example
-       */
-      clearViewport(): Chainable;
-    }
-  }
-}
-
-Cypress.Commands.add("clearViewport", () => {
-  const runnerContainer =
-    window.parent.document.getElementsByClassName("iframes-container")[0];
-  runnerContainer.setAttribute(
-    "style",
-    "left: 0; top: 0; width: 100%; height: 100%;"
-  );
-
-  const sizeContainer =
-    window.parent.document.getElementsByClassName("size-container")[0];
-  sizeContainer.setAttribute("style", "");
-
-  const sidebar =
-    window.parent.document.getElementsByClassName("reporter-wrap")[0];
-  sidebar.setAttribute("style", "opacity: 0");
-
-  const header = window.parent.document.querySelector(
-    ".runner.container header"
-  );
-  header.setAttribute("style", "opacity: 0");
-
-  return runnerContainer;
-});
